Add major cleanup and refactoring of rust module
